refactor: share WeatherNext 2 base
