+++ earth2studio/models/px/__init__.py
@@ -55,8 +55,10 @@ from earth2studio.models.px.stormscope import (
 )
 from earth2studio.models.px.stormscope_meteosat import StormScopeMeteosatEU
 from earth2studio.models.px.ucast import UCast
-from earth2studio.models.px.weathernext2_cyclones import WeatherNext2Cyclones
-from earth2studio.models.px.weathernext2_cyclones_mini import WeatherNext2CyclonesMini
+from earth2studio.models.px.weathernext2_cyclones import (
+    WeatherNext2Cyclones,
+    WeatherNext2CyclonesMini,
+)
 
 # Silence warning spam from various models
 warnings.filterwarnings("ignore")
